Made some cleaning on Dev branch
