--- help_env.c
@@ -1,18 +1,0 @@
-#include "shell.h"
-
-/**
- * help_env - An helper function to display env help page
- */
-
-void help_env(void)
-{
-	printf("⌌--------------🐱--------------⌍\n");
-	printf("|           Help env           |\n");
-	printf("|------------------------------|\n");
-	printf("| env: This command will       |\n");
-	printf("| display the current          |\n");
-	printf("| environment                  |\n");
-	printf("|--------------😎--------------|\n");
-	printf("| usage : env                  |\n");
-	printf("⌎------------------------------⌏\n");
-}
