--- help_exit.c
@@ -1,18 +1,0 @@
-#include "shell.h"
-
-/**
- * help_exit - Helper function to display exit help page
- */
-
-void help_exit(void)
-{
-	printf("⌌--------------🐱--------------⌍\n");
-	printf("|           Help exit          |\n");
-	printf("|------------------------------|\n");
-	printf("| exit: This command will      |\n");
-	printf("| exit the shell with the      |\n");
-	printf("| given exit status            |\n");
-	printf("|--------------😎--------------|\n");
-	printf("| usage : exit [status]        |\n");
-	printf("⌎------------------------------⌏\n");
-}
