--- help_help.c
@@ -1,18 +1,0 @@
-#include "shell.h"
-
-/**
- * help_help - Helper function to display help help page
- */
-
-void help_help(void)
-{
-	printf("⌌--------------🐱--------------⌍\n");
-	printf("|           Help help          |\n");
-	printf("|------------------------------|\n");
-	printf("| help: This command will      |\n");
-	printf("| display the help page        |\n");
-	printf("| in the shell                 |\n");
-	printf("|--------------😎--------------|\n");
-	printf("| usage : help [BUILTIN]       |\n");
-	printf("⌎------------------------------⌏\n");
-}
